Modify Verification Binary File

+++ src/main/java/oss/fosslight/service/impl/VerificationServiceImpl.java
@@ -1406,10 +1406,10 @@ public class VerificationServiceImpl extends CoTopComponent implements Verificat
 			if (f.exists()) {
 				boolean isExistFile = false;
 				for (File bFile : f.listFiles()){
-					if (bFile.exists()) {
+					if (bFile.exists() &&f.listFiles().length == 1) {
 						isExistFile = true;
-						File copiedFile = new File(VERIFY_PATH_OUTPUT +"/"+prjId+"/verify_binary_list");
-						if (!copiedFile.exists()) {
+						if (packagingFileIdx == 1) {
+							File copiedFile = new File(VERIFY_PATH_OUTPUT + "/" + prjId + "/verify_binary_list");
 							Files.copy(bFile.toPath(), copiedFile.toPath(), StandardCopyOption.REPLACE_EXISTING);
 						} else {
 							FileUtil.addFileContents(VERIFY_PATH_OUTPUT +"/"+prjId+"/verify_binary_list", bFile.getPath());
